--- a/library/src/common/rocsolver_logger.cpp
+++ b/library/src/common/rocsolver_logger.cpp
@@ -1,5 +1,5 @@
 /* **************************************************************************
- * Copyright (C) 2021-2024 Advanced Micro Devices, Inc. All rights reserved.
+ * Copyright (C) 2021-2025 Advanced Micro Devices, Inc. All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
@@ -324,7 +324,7 @@ void rocsolver_logger::accumulate_times(rocsolver_profile_map& m)
             if (err != hipSuccess)
                 printf("BADNESS IN ELAPSED TIME\n");
                 printf("error: %s\n", hipGetErrorString(err));
-            entry.total_time += ms * 1000; // µs
+            entry.total_time += ms * 1000; // us
             // Destroy events after measurement
             err = hipEventDestroy(pair.first);
             if (err != hipSuccess)