… library/src/CMakeLists.txt
… library/src/common/rocsolver_logger.cpp.old
… library/src/include/rocsolver_logger.hpp.old
… library/src/lapack/roclapack_geqr2.hpp
… library/src/lapack/roclapack_gesdd.hpp
… library/src/lapack/roclapack_sytd2_hetd2.hpp
… library/src/specialized/roclapack_gemm_specialized_kernels.hpp