Currently the rocSOLVER logger used a synchronizing method after every kernel launch (when --profile or --profile-kernels is enabled in rocsolver-bench) to record the time elapsed inside the kernel. This makes the benchmarking tool slower, and provides inaccurate timing results for methods (especially those that launch a lot of kernels).

This PR uses hipEvents, and a single synchronization at the end of the top level function call. This reduces the overhead by the logger. 

Before:
```
johtyler@DOCKER-cgy-decidueye:~/code/rocSOLVER/build/release/clients/staging$ ./rocsolver-bench -f syevd --evect V -n 1024 -r s --profile 5 --profile_kernels 1
rocSOLVER version 3.31.0.e40d6537-dirty (with rocBLAS 5.0.0.ce2f7eba97)
Query device success: there are 1 devices
-------------------------------------------------------------------------------
Device ID 0 : AMD Radeon RX 6900 XT
with 17.2 GB memory, max. SCLK 2660 MHz, max. MCLK 1000 MHz, compute capability 10.3
maxGridDimX 2147483647, sharedMemPerBlock 65.5 KB, maxThreadsPerBlock 1024, warpSize 32
-------------------------------------------------------------------------------

============================================
Arguments:
============================================
evect           uplo            n               lda
V               U               1024            1024

============================================
Results:
============================================
cpu_time_us     gpu_time_us
908987          129632.1

------- PROFILE -------
rocsolver_syevd_heevd_template: Calls: 10, Total Time: 1296.073 ms (in nested functions: 1295.909 ms)
    copy_mat<T>: Calls: 10, Total Time: 0.253 ms
    rocsolver_stedc_template: Calls: 10, Total Time: 71.499 ms (in nested functions: 71.160 ms)
        stedc_copyD: Calls: 10, Total Time: 0.172 ms
        (stedc_mergeUpdate_kernel<S>): Calls: 60, Total Time: 1.281 ms
        rocsolver_gemm_template: Calls: 60, Total Time: 9.171 ms (in nested functions: 9.094 ms)
            rocblas_gemm_template: Calls: 60, Total Time: 9.094 ms
        (stedc_mergePrepare_DeflateCount_kernel<S>): Calls: 60, Total Time: 1.029 ms
        init_ident<S>: Calls: 10, Total Time: 0.344 ms
        (stedc_divide_kernel<S>): Calls: 10, Total Time: 0.458 ms
        (stedc_mergePrepare_DeflateApply_kernel<S>): Calls: 60, Total Time: 1.047 ms
        stedc_reshuffleC<S>: Calls: 60, Total Time: 1.549 ms
        stedc_update_splits: Calls: 60, Total Time: 1.073 ms
        (stedc_mergePrepare_DeflateZero_kernel<S>): Calls: 60, Total Time: 1.174 ms
        stedc_copyC<T>: Calls: 10, Total Time: 0.248 ms
        (stedc_solve_kernel<S>): Calls: 10, Total Time: 2.036 ms
        (stedc_mergePrepare_SortD_kernel<S>): Calls: 60, Total Time: 1.679 ms
        (stedc_mergeValues_Rescale_kernel<S>): Calls: 60, Total Time: 3.708 ms
        reset_info: Calls: 10, Total Time: 0.165 ms
        (stedc_mergePrepare_SetCandFlags_kernel<S>): Calls: 60, Total Time: 0.990 ms
        stedc_sort<T>: Calls: 10, Total Time: 0.351 ms
        (stedc_mergeVectors_kernel<STEDC_EXTERNAL_GEMM, S>): Calls: 60, Total Time: 1.864 ms
        stedc_copyC<S>: Calls: 60, Total Time: 1.500 ms
        (stedc_mergeRotate_kernel<S>): Calls: 60, Total Time: 1.204 ms
        (stedc_mergeValues_SortDZ_kernel<S>): Calls: 60, Total Time: 1.773 ms
        (stedc_mergeValues_Solve_kernel<S>): Calls: 60, Total Time: 37.145 ms
        (stedc_mergeValues_copyD_kernel<S>): Calls: 60, Total Time: 1.199 ms
    rocsolver_sytrd_hetrd_template: Calls: 10, Total Time: 1165.915 ms (in nested functions: 1165.748 ms)
        rocsolver_sytd2_hetd2_template: Calls: 10, Total Time: 183.312 ms (in nested functions: 181.326 ms)
            set_tridiag<T>: Calls: 10, Total Time: 0.178 ms
            (sytd2_upper_kernel_small<256, T>): Calls: 10, Total Time: 7.085 ms
            rocblas_syr2_template: Calls: 1300, Total Time: 24.354 ms
            set_tau<T>: Calls: 1300, Total Time: 20.889 ms
            (latrd_dot_scale_axpy<64, T>): Calls: 1300, Total Time: 23.332 ms
            rocblas_symv_template: Calls: 1300, Total Time: 35.834 ms
            rocsolver_larfg_template: Calls: 1300, Total Time: 69.654 ms (in nested functions: 66.514 ms)
                rocblas_scal_template: Calls: 1300, Total Time: 21.727 ms
                (set_taubeta<T, I>): Calls: 1300, Total Time: 21.972 ms
                rocblas_dot_template: Calls: 1300, Total Time: 22.815 ms
        rocsolver_gemm_template: Calls: 240, Total Time: 7.915 ms (in nested functions: 7.659 ms)
            rocblas_gemm_template: Calls: 240, Total Time: 7.659 ms
        set_tridiag<T>: Calls: 10, Total Time: 0.173 ms
        rocsolver_latrd_template: Calls: 120, Total Time: 973.124 ms (in nested functions: 962.227 ms)
            (latrd_dot_scale_axpy<1024, T>): Calls: 7680, Total Time: 130.341 ms
            latrd_upper_updateW_kernel<T>: Calls: 7680, Total Time: 137.395 ms
            rocsolver_larfg_template: Calls: 7680, Total Time: 405.339 ms (in nested functions: 387.903 ms)
                rocblas_scal_template: Calls: 7680, Total Time: 127.294 ms
                (set_taubeta<T, I>): Calls: 7680, Total Time: 128.847 ms
                rocblas_dot_template: Calls: 7680, Total Time: 131.762 ms
            (latrd_upper_computeW_gemvt_kernel<NB, T>): Calls: 7680, Total Time: 155.883 ms
            latrd_upper_updateA_kernel<T>: Calls: 7680, Total Time: 133.269 ms
        (copy_trans_mat<T, T>): Calls: 10, Total Time: 1.224 ms
    rocsolver_ormtr_unmtr_template: Calls: 10, Total Time: 57.771 ms (in nested functions: 57.761 ms)
        rocsolver_ormql_unmql_template: Calls: 10, Total Time: 57.761 ms (in nested functions: 57.592 ms)
            rocsolver_larfb_inverse_template: Calls: 160, Total Time: 41.433 ms (in nested functions: 40.825 ms)
                rocsolver_gemm_template: Calls: 300, Total Time: 12.643 ms (in nested functions: 12.133 ms)
                addmatA1: Calls: 160, Total Time: 7.499 ms
                rocsolver_trsm_lower_template: Calls: 160, Total Time: 7.800 ms (in nested functions: 7.576 ms)
                rocblas_trmm_template: Calls: 320, Total Time: 9.079 ms
                copymatA1: Calls: 160, Total Time: 3.804 ms
            rocsolver_larft_inverse_template: Calls: 160, Total Time: 16.159 ms (in nested functions: 15.803 ms)
                (larft_restore_tri): Calls: 160, Total Time: 2.612 ms
                larft_set_diag: Calls: 160, Total Time: 2.680 ms
                rocsolver_gemm_template: Calls: 160, Total Time: 7.829 ms (in nested functions: 7.593 ms)
                (larft_set_tri): Calls: 160, Total Time: 2.682 ms
    reset_info: Calls: 10, Total Time: 0.471 ms
iota_n<T>: Calls: 10, Total Time: 0.168 ms
```

After:
```
johtyler@DOCKER-cgy-decidueye:~/code/rocSOLVER/build/release/clients/staging$ ./rocsolver-bench -f syevd --evect V -n 1024 -r s --profile 15 --profile_kernels 1
rocSOLVER version 3.31.0.e40d6537 (with rocBLAS 5.0.0.ce2f7eba97)
Query device success: there are 1 devices
-------------------------------------------------------------------------------
Device ID 0 : AMD Radeon RX 6900 XT
with 17.2 GB memory, max. SCLK 2660 MHz, max. MCLK 1000 MHz, compute capability 10.3
maxGridDimX 2147483647, sharedMemPerBlock 65.5 KB, maxThreadsPerBlock 1024, warpSize 32
-------------------------------------------------------------------------------

============================================
Arguments:
============================================
evect           uplo            n               lda
V               U               1024            1024

============================================
Results:
============================================
cpu_time_us     gpu_time_us
904202          110882.1

------- PROFILE -------
rocsolver_syevd_heevd_template: Calls: 10, Total Time: 1108.372 ms (in nested functions: 1107.678 ms)
    copy_mat<T>: Calls: 10, Total Time: 0.168 ms
    rocsolver_stedc_template: Calls: 10, Total Time: 66.820 ms (in nested functions: 61.878 ms)
        stedc_copyD: Calls: 10, Total Time: 0.092 ms
        (stedc_mergeUpdate_kernel<S>): Calls: 60, Total Time: 0.767 ms
        rocsolver_gemm_template: Calls: 60, Total Time: 8.675 ms (in nested functions: 8.121 ms)
            rocblas_gemm_template: Calls: 60, Total Time: 8.121 ms
        (stedc_mergePrepare_DeflateCount_kernel<S>): Calls: 60, Total Time: 0.534 ms
        init_ident<S>: Calls: 10, Total Time: 0.160 ms
        (stedc_divide_kernel<S>): Calls: 10, Total Time: 0.362 ms
        (stedc_mergePrepare_DeflateApply_kernel<S>): Calls: 60, Total Time: 0.538 ms
        stedc_reshuffleC<S>: Calls: 60, Total Time: 1.011 ms
        stedc_update_splits: Calls: 60, Total Time: 0.594 ms
        (stedc_mergePrepare_DeflateZero_kernel<S>): Calls: 60, Total Time: 0.660 ms
        stedc_copyC<T>: Calls: 10, Total Time: 0.166 ms
        (stedc_solve_kernel<S>): Calls: 10, Total Time: 1.944 ms
        (stedc_mergePrepare_SortD_kernel<S>): Calls: 60, Total Time: 1.251 ms
        (stedc_mergeValues_Rescale_kernel<S>): Calls: 60, Total Time: 2.831 ms
        reset_info: Calls: 10, Total Time: 0.081 ms
        (stedc_mergePrepare_SetCandFlags_kernel<S>): Calls: 60, Total Time: 0.495 ms
        stedc_sort<T>: Calls: 10, Total Time: 0.269 ms
        (stedc_mergeVectors_kernel<STEDC_EXTERNAL_GEMM, S>): Calls: 60, Total Time: 1.318 ms
        stedc_copyC<S>: Calls: 60, Total Time: 0.963 ms
        (stedc_mergeRotate_kernel<S>): Calls: 60, Total Time: 0.683 ms
        (stedc_mergeValues_SortDZ_kernel<S>): Calls: 60, Total Time: 1.297 ms
        (stedc_mergeValues_Solve_kernel<S>): Calls: 60, Total Time: 36.519 ms
        (stedc_mergeValues_copyD_kernel<S>): Calls: 60, Total Time: 0.666 ms
    rocsolver_sytrd_hetrd_template: Calls: 10, Total Time: 984.667 ms (in nested functions: 982.796 ms)
        rocsolver_sytd2_hetd2_template: Calls: 10, Total Time: 158.080 ms (in nested functions: 126.732 ms)
            set_tridiag<T>: Calls: 10, Total Time: 0.089 ms
            (sytd2_upper_kernel_small<256, T>): Calls: 10, Total Time: 7.000 ms
            rocblas_syr2_template: Calls: 1300, Total Time: 13.256 ms
            set_tau<T>: Calls: 1300, Total Time: 10.693 ms
            (latrd_dot_scale_axpy<64, T>): Calls: 1300, Total Time: 12.558 ms
            rocblas_symv_template: Calls: 1300, Total Time: 25.007 ms
            rocsolver_larfg_template: Calls: 1300, Total Time: 58.129 ms (in nested functions: 33.155 ms)
                rocblas_scal_template: Calls: 1300, Total Time: 10.950 ms
                (set_taubeta<T, I>): Calls: 1300, Total Time: 11.151 ms
                rocblas_dot_template: Calls: 1300, Total Time: 11.054 ms
        rocsolver_gemm_template: Calls: 240, Total Time: 6.420 ms (in nested functions: 4.172 ms)
            rocblas_gemm_template: Calls: 240, Total Time: 4.172 ms
        set_tridiag<T>: Calls: 10, Total Time: 0.090 ms
        rocsolver_latrd_template: Calls: 120, Total Time: 817.062 ms (in nested functions: 636.445 ms)
            (latrd_dot_scale_axpy<1024, T>): Calls: 7680, Total Time: 66.671 ms
            latrd_upper_updateW_kernel<T>: Calls: 7680, Total Time: 71.004 ms
            rocsolver_larfg_template: Calls: 7680, Total Time: 336.354 ms (in nested functions: 192.328 ms)
                rocblas_scal_template: Calls: 7680, Total Time: 63.471 ms
                (set_taubeta<T, I>): Calls: 7680, Total Time: 64.289 ms
                rocblas_dot_template: Calls: 7680, Total Time: 64.568 ms
            (latrd_upper_computeW_gemvt_kernel<NB, T>): Calls: 7680, Total Time: 92.295 ms
            latrd_upper_updateA_kernel<T>: Calls: 7680, Total Time: 70.123 ms
        (copy_trans_mat<T, T>): Calls: 10, Total Time: 1.144 ms
    rocsolver_ormtr_unmtr_template: Calls: 10, Total Time: 55.944 ms (in nested functions: 55.851 ms)
        rocsolver_ormql_unmql_template: Calls: 10, Total Time: 55.851 ms (in nested functions: 54.284 ms)
            rocsolver_larfb_inverse_template: Calls: 160, Total Time: 39.484 ms (in nested functions: 33.509 ms)
                rocsolver_gemm_template: Calls: 300, Total Time: 11.120 ms (in nested functions: 8.270 ms)
                    rocblas_gemm_template: Calls: 300, Total Time: 8.270 ms
                addmatA1: Calls: 160, Total Time: 6.040 ms
                rocsolver_trsm_lower_template: Calls: 160, Total Time: 7.671 ms (in nested functions: 6.156 ms)
                    nonunit_forward_substitution_kernel<T>: Calls: 160, Total Time: 6.156 ms
                rocblas_trmm_template: Calls: 320, Total Time: 6.194 ms
                copymatA1: Calls: 160, Total Time: 2.484 ms
            rocsolver_larft_inverse_template: Calls: 160, Total Time: 14.800 ms (in nested functions: 10.989 ms)
                (larft_restore_tri): Calls: 160, Total Time: 1.330 ms
                larft_set_diag: Calls: 160, Total Time: 1.312 ms
                rocsolver_gemm_template: Calls: 160, Total Time: 7.008 ms (in nested functions: 5.491 ms)
                    rocblas_gemm_template: Calls: 160, Total Time: 5.491 ms
                (larft_set_tri): Calls: 160, Total Time: 1.340 ms
    reset_info: Calls: 10, Total Time: 0.078 ms
iota_n<T>: Calls: 10, Total Time: 0.080 ms

```

However, for larger matrix sizes, the difference does not seem to be as large:
Before:
```
johtyler@DOCKER-cgy-decidueye:~/code/rocSOLVER/build/release/clients/staging$ ./rocsolver-bench -f syevd --evect V -n 8192 -r s --profile 5 --profile_kernels 1
rocSOLVER version 3.31.0.e40d6537-dirty (with rocBLAS 5.0.0.ce2f7eba97)
Query device success: there are 1 devices
-------------------------------------------------------------------------------
Device ID 0 : AMD Radeon RX 6900 XT
with 17.2 GB memory, max. SCLK 2660 MHz, max. MCLK 1000 MHz, compute capability 10.3
maxGridDimX 2147483647, sharedMemPerBlock 65.5 KB, maxThreadsPerBlock 1024, warpSize 32
-------------------------------------------------------------------------------

============================================
Arguments:
============================================
evect           uplo            n               lda
V               U               8192            8192

============================================
Results:
============================================
cpu_time_us     gpu_time_us
446644674       3395728.1

------- PROFILE -------
rocsolver_syevd_heevd_template: Calls: 10, Total Time: 33956.892 ms (in nested functions: 33956.608 ms)
    copy_mat<T>: Calls: 10, Total Time: 15.541 ms
    rocsolver_stedc_template: Calls: 10, Total Time: 2383.470 ms (in nested functions: 2382.838 ms)
        stedc_copyD: Calls: 10, Total Time: 0.192 ms
        (stedc_mergeUpdate_kernel<S>): Calls: 90, Total Time: 22.789 ms
        rocsolver_gemm_template: Calls: 90, Total Time: 738.724 ms (in nested functions: 738.625 ms)
            rocblas_gemm_template: Calls: 90, Total Time: 738.625 ms
        (stedc_mergePrepare_DeflateCount_kernel<S>): Calls: 90, Total Time: 1.619 ms
        init_ident<S>: Calls: 10, Total Time: 22.444 ms
        (stedc_divide_kernel<S>): Calls: 10, Total Time: 2.104 ms
        (stedc_mergePrepare_DeflateApply_kernel<S>): Calls: 90, Total Time: 1.681 ms
        stedc_reshuffleC<S>: Calls: 90, Total Time: 125.818 ms
        stedc_update_splits: Calls: 90, Total Time: 2.388 ms
        (stedc_mergePrepare_DeflateZero_kernel<S>): Calls: 90, Total Time: 3.599 ms
        stedc_copyC<T>: Calls: 10, Total Time: 13.534 ms
        (stedc_solve_kernel<S>): Calls: 10, Total Time: 2.507 ms
        (stedc_mergePrepare_SortD_kernel<S>): Calls: 90, Total Time: 10.748 ms
        (stedc_mergeValues_Rescale_kernel<S>): Calls: 90, Total Time: 308.987 ms
        reset_info: Calls: 10, Total Time: 0.178 ms
        (stedc_mergePrepare_SetCandFlags_kernel<S>): Calls: 90, Total Time: 1.534 ms
        stedc_sort<T>: Calls: 10, Total Time: 13.548 ms
        (stedc_mergeVectors_kernel<STEDC_EXTERNAL_GEMM, S>): Calls: 90, Total Time: 64.841 ms
        stedc_copyC<S>: Calls: 90, Total Time: 143.988 ms
        (stedc_mergeRotate_kernel<S>): Calls: 90, Total Time: 4.818 ms
        (stedc_mergeValues_SortDZ_kernel<S>): Calls: 90, Total Time: 11.464 ms
        (stedc_mergeValues_Solve_kernel<S>): Calls: 90, Total Time: 869.310 ms
        (stedc_mergeValues_copyD_kernel<S>): Calls: 90, Total Time: 16.023 ms
    rocsolver_sytrd_hetrd_template: Calls: 10, Total Time: 26257.438 ms (in nested functions: 26255.980 ms)
        rocsolver_sytd2_hetd2_template: Calls: 10, Total Time: 189.785 ms (in nested functions: 187.988 ms)
            set_tridiag<T>: Calls: 10, Total Time: 0.184 ms
            (sytd2_upper_kernel_small<256, T>): Calls: 10, Total Time: 7.149 ms
            rocblas_syr2_template: Calls: 1300, Total Time: 24.910 ms
            set_tau<T>: Calls: 1300, Total Time: 21.162 ms
            (latrd_dot_scale_axpy<64, T>): Calls: 1300, Total Time: 24.077 ms
            rocblas_symv_template: Calls: 1300, Total Time: 37.615 ms
            rocsolver_larfg_template: Calls: 1300, Total Time: 72.891 ms (in nested functions: 70.156 ms)
                rocblas_scal_template: Calls: 1300, Total Time: 22.670 ms
                (set_taubeta<T, I>): Calls: 1300, Total Time: 23.637 ms
                rocblas_dot_template: Calls: 1300, Total Time: 23.849 ms
        rocsolver_gemm_template: Calls: 2480, Total Time: 1037.996 ms (in nested functions: 1035.631 ms)
            rocblas_gemm_template: Calls: 2480, Total Time: 1035.631 ms
        set_tridiag<T>: Calls: 10, Total Time: 0.206 ms
        rocsolver_latrd_template: Calls: 1240, Total Time: 24255.211 ms (in nested functions: 24146.721 ms)
            (latrd_dot_scale_axpy<1024, T>): Calls: 79360, Total Time: 1435.602 ms
            latrd_upper_updateW_kernel<T>: Calls: 79360, Total Time: 1558.961 ms
            rocsolver_larfg_template: Calls: 79360, Total Time: 4273.121 ms (in nested functions: 4099.070 ms)
                rocblas_scal_template: Calls: 79360, Total Time: 1305.078 ms
                (set_taubeta<T, I>): Calls: 79360, Total Time: 1393.120 ms
                rocblas_dot_template: Calls: 79360, Total Time: 1400.872 ms
            (latrd_upper_computeW_gemvt_kernel<NB, T>): Calls: 79360, Total Time: 15429.843 ms
            latrd_upper_updateA_kernel<T>: Calls: 79360, Total Time: 1449.194 ms
        (copy_trans_mat<T, T>): Calls: 10, Total Time: 772.782 ms
    rocsolver_ormtr_unmtr_template: Calls: 10, Total Time: 5275.564 ms (in nested functions: 5275.557 ms)
        rocsolver_ormql_unmql_template: Calls: 10, Total Time: 5275.557 ms (in nested functions: 5274.601 ms)
            rocsolver_larfb_inverse_template: Calls: 1280, Total Time: 4869.016 ms (in nested functions: 4864.104 ms)
                rocsolver_gemm_template: Calls: 2540, Total Time: 3737.417 ms (in nested functions: 3733.820 ms)
                addmatA1: Calls: 1280, Total Time: 505.159 ms
                rocsolver_trsm_lower_template: Calls: 1280, Total Time: 253.087 ms (in nested functions: 249.506 ms)
                rocblas_trmm_template: Calls: 2560, Total Time: 159.810 ms
                copymatA1: Calls: 1280, Total Time: 208.631 ms
            rocsolver_larft_inverse_template: Calls: 1280, Total Time: 405.585 ms (in nested functions: 402.618 ms)
                (larft_restore_tri): Calls: 1280, Total Time: 21.990 ms
                larft_set_diag: Calls: 1280, Total Time: 21.910 ms
                rocsolver_gemm_template: Calls: 1280, Total Time: 336.110 ms (in nested functions: 334.354 ms)
                (larft_set_tri): Calls: 1280, Total Time: 22.608 ms
    reset_info: Calls: 10, Total Time: 24.595 ms
iota_n<T>: Calls: 10, Total Time: 0.177 ms
```

After:
```
…